global exception handling, and service/controller improvements

+++ src/main/java/bd/course/work/repositories/QuestRepository.java
@@ -3,6 +3,6 @@ package bd.course.work.repositories;
 import bd.course.work.entities.Quest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestRepository extends JpaRepository<Quest, Integer> {
+public interface QuestRepository extends JpaRepository<Quest, Long> {
 
 }
